software/requirements.txt
langchain
langchain-openai
langfuse>=3,<4
python-dotenv
ulid-py
pandas
python-dotenv
python=3.13
